feat: add effect fadeIn/fadeOut on card
